All tests now pass again.
